009 - Optical Character Recognition/pyproject.toml
[tool.poetry]
name = "Labo3"
version = "0.1.0"
description = ""
authors = ["sergio.guarino <sergio.guarino@etu.hesge.ch>"]

[tool.poetry.dependencies]
python = "^3.8"
tensorflow = "^2.5.0"
black = "^21.5b1"
mypy = "^0.812"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
aiofiles = "^0.7.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
